+++ supervisor/create_windows.go
@@ -1,0 +1,16 @@
+package supervisor
+
+type StartTask struct {
+	baseTask
+	ID            string
+	BundlePath    string
+	Stdout        string
+	Stderr        string
+	Stdin         string
+	StartResponse chan StartResponse
+	Labels        []string
+}
+
+// Checkpoint not supported on Windows
+func setTaskCheckpoint(t *StartTask) {
+}
